Refactored ketchup-radio to kup-radio
