+++ packages/ketchup/src/components/kup-radio/readme.md
@@ -1,4 +1,4 @@
-# ketchup-radio
+# kup-radio
