+++ packages/ketchup/src/index.html
@@ -65,11 +65,11 @@
         <a href="data-table.html">Ketchup data table</a>
         <br />
         <kup-button label="Sono bottone"></kup-button>
-        <ketchup-radio
+        <kup-radio
             id="krg1"
             displayed-field="label"
             value-field="value"
-        ></ketchup-radio>
+        ></kup-radio>
         <ketchup-text-input
             id="int1"
             is-clearable
